requirements.txt
# pip install --target=lib
google-endpoints
Flask
